+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.5.0] - 2021-04-30
+
 -   External dependencies can be added as optionals in configuration
 -   Support for labels, tolerations and annotations in k8s
 -   Logging added to operators
@@ -63,7 +65,9 @@
 
 -   Initial implementation of `kedro_airflow_k8s` plugin stub. 
 
-[Unreleased]: https://github.com/getindata/kedro-airflow-k8s/compare/0.4.0...HEAD
+[Unreleased]: https://github.com/getindata/kedro-airflow-k8s/compare/0.5.0...HEAD
+
+[0.5.0]: https://github.com/getindata/kedro-airflow-k8s/compare/0.4.0...0.5.0
 
 [0.4.0]: https://github.com/getindata/kedro-airflow-k8s/compare/0.3.0...0.4.0